+++ inst/assets/buttons-group/radioGroupButtons-bindings.js
@@ -17,7 +17,7 @@
 var radioGroupButtonsBinding = new Shiny.InputBinding();
 $.extend(radioGroupButtonsBinding, {
   find: function(scope) {
-    return $(scope).find(".radioGroupButtons");
+    return $(scope).find(".radio-group-buttons");
   },
   getId: function(el) {
     return el.id;
@@ -103,19 +103,25 @@ $.extend(radioGroupButtonsBinding, {
     if (data.disabled) {
       $el.find("button").attr("disabled", "disabled");
       $el.find("button").addClass("disabled");
+      $el.find("label").addClass("disabled");
     } else {
       $el.find("button").removeAttr("disabled");
       $el.find("button").removeClass("disabled");
+      $el.find("label").removeClass("disabled");
     }
     if (data.hasOwnProperty("disabledChoices")) {
       for (var i = 0; i < data.disabledChoices.length; i++) {
-        $(
+        var toDisable = $(
           'input:radio[name="' +
             Shiny.$escape(el.id) +
             '"][value="' +
             Shiny.$escape(data.disabledChoices[i]) +
             '"]'
-        )
+        );
+        toDisable
+          .next("label")
+          .addClass("disabled");
+        toDisable
           .parent()
           .attr("disabled", "disabled")
           .addClass("disabled");